+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2025-01-23
+### Changed
+- Made the client compatible also with older versions of Jackson like version 2.8.8
+
+### Added
+- Examples for Google Gemini model
+
 ## [0.2.1] - 2025-01-10
 ### Changed
 - streamChat with new implementation to accumulate the chunks
 
 ## [0.2.0] - 2025-01-06
-### added
+### Added
 - Initial release of the project
 
 ## [0.1.0] - 2024-12-20